+++ src/tests/cell.cpp
@@ -162,6 +162,7 @@ TEST_CASE("Cell") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto np = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != np);
@@ -201,6 +202,7 @@ TEST_CASE("Cell") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto np = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != np);
@@ -240,6 +242,7 @@ TEST_CASE("Cell") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto np = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != np);
@@ -279,6 +282,7 @@ TEST_CASE("Cell") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto np = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != np);
@@ -319,6 +323,7 @@ TEST_CASE("Cell") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto np = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != np);