+++ src/tests/geom.cpp
@@ -33,6 +33,7 @@ TEST_CASE("Geometry") {
         .name = nullptr,
         .resizecb = nullptr,
         .flags = 0,
+        .margin_b = 0, .margin_r = 0,
       };
       auto n = ncplane_create(n_, &nopts);
       REQUIRE(n);
@@ -79,6 +80,7 @@ TEST_CASE("Geometry") {
         .name = nullptr,
         .resizecb = nullptr,
         .flags = 0,
+        .margin_b = 0, .margin_r = 0,
       };
       auto n = ncplane_create(n_, &nopts);
       REQUIRE(n);