+++ src/tests/layout.cpp
@@ -21,6 +21,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -45,6 +46,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -69,6 +71,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -94,6 +97,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -120,6 +124,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -146,6 +151,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -172,6 +178,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -199,6 +206,7 @@ TEST_CASE("TextLayout") {
         .name = nullptr,
         .resizecb = nullptr,
         .flags = 0,
+        .margin_b = 0, .margin_r = 0,
       };
       auto sp = ncplane_create(n_, &nopts);
       REQUIRE(sp);
@@ -227,6 +235,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -255,6 +264,7 @@ TEST_CASE("TextLayout") {
         .name = nullptr,
         .resizecb = nullptr,
         .flags = 0,
+        .margin_b = 0, .margin_r = 0,
       };
       auto sp = ncplane_create(n_, &nopts);
       REQUIRE(sp);
@@ -281,6 +291,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -307,6 +318,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -333,6 +345,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -359,6 +372,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -385,6 +399,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -411,6 +426,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -438,6 +454,7 @@ TEST_CASE("TextLayout") {
         .name = nullptr,
         .resizecb = nullptr,
         .flags = 0,
+        .margin_b = 0, .margin_r = 0,
       };
       auto sp = ncplane_create(n_, &nopts);
       REQUIRE(sp);
@@ -466,6 +483,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -492,6 +510,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -526,6 +545,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -560,6 +580,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);
@@ -597,6 +618,7 @@ TEST_CASE("TextLayout") {
       .name = nullptr,
       .resizecb = nullptr,
       .flags = 0,
+      .margin_b = 0, .margin_r = 0,
     };
     auto sp = ncplane_create(n_, &nopts);
     REQUIRE(sp);